+++ tests/collections/asr/decoding/conftest.py
@@ -17,6 +17,7 @@ import pytest
 
 from nemo.collections.asr.models import ASRModel
 from nemo.collections.asr.parts.utils.manifest_utils import read_manifest, write_manifest
+from tests.collections.asr.decoding.utils import make_preprocessor_deterministic, preserve_decoding_cfg_and_cpu_device
 
 CHECKPOINTS_PATH = Path("/home/TestData/asr")
 
@@ -55,8 +56,8 @@ def an4_train_manifest_corrected(tmp_path_factory, test_data_dir):
     return an4_train_manifest_corrected_path
 
 
-@pytest.fixture(scope="module")
-def stt_en_fastconformer_transducer_large():
+@pytest.fixture(scope="package")
+def _stt_en_fastconformer_transducer_large_raw():
     if CHECKPOINTS_PATH.exists():
         model = ASRModel.restore_from(
             str(CHECKPOINTS_PATH / "stt_en_fastconformer_transducer_large.nemo"), map_location="cpu"
@@ -64,11 +65,12 @@ def stt_en_fastconformer_transducer_large():
     else:
         model_name = "stt_en_fastconformer_transducer_large"
         model = ASRModel.from_pretrained(model_name, map_location="cpu")
-    return model.eval()
+    make_preprocessor_deterministic(model)
+    return model
 
 
-@pytest.fixture(scope="module")
-def stt_en_fastconformer_tdt_large():
+@pytest.fixture(scope="package")
+def _stt_en_fastconformer_tdt_large_raw():
     if CHECKPOINTS_PATH.exists():
         model = ASRModel.restore_from(
             str(CHECKPOINTS_PATH / "stt_en_fastconformer_tdt_large.nemo"), map_location="cpu"
@@ -76,4 +78,19 @@ def stt_en_fastconformer_tdt_large():
     else:
         model_name = "nvidia/stt_en_fastconformer_tdt_large"
         model = ASRModel.from_pretrained(model_name, map_location="cpu")
-    return model.eval()
+    make_preprocessor_deterministic(model)
+    return model
+
+
+@pytest.fixture
+def stt_en_fastconformer_transducer_large(_stt_en_fastconformer_transducer_large_raw):
+    model = _stt_en_fastconformer_transducer_large_raw
+    with preserve_decoding_cfg_and_cpu_device(model):
+        yield model
+
+
+@pytest.fixture
+def stt_en_fastconformer_tdt_large(_stt_en_fastconformer_tdt_large_raw):
+    model = _stt_en_fastconformer_tdt_large_raw
+    with preserve_decoding_cfg_and_cpu_device(model):
+        yield model
