+++ tests/collections/asr/decoding/test_batched_hyps_and_alignments.py
@@ -19,18 +19,7 @@ import pytest
 import torch
 
 from nemo.collections.asr.parts.utils.rnnt_utils import BatchedAlignments, BatchedHyps, batched_hyps_to_hypotheses
-
-
-@contextmanager
-def avoid_sync_operations(device: torch.device):
-    try:
-        if device.type == "cuda":
-            torch.cuda.set_sync_debug_mode(2)  # fail if a blocking operation
-        yield
-    finally:
-        if device.type == "cuda":
-            torch.cuda.set_sync_debug_mode(0)  # default, blocking operations are allowed
-
+from tests.collections.asr.decoding.utils import avoid_sync_operations
 
 DEVICES: List[torch.device] = [torch.device("cpu")]
